+++ Cargo.toml
@@ -8,3 +8,8 @@ members = [
     "cip25-wasm",
     "wasm"
 ]
+
+# exclude old crate structure to avoid error in it
+exclude = [
+    "rust"
+]
