+++ core/src/lib.rs
@@ -5,6 +5,3 @@ pub mod serialization;
 
 pub mod ordered_hash_map;
 
-use ordered_hash_map::OrderedHashMap;
-
-use cbor_event::{Sz, LenSz, StringLenSz};
