+++ crypto-wasm/README.md
@@ -1,0 +1,3 @@
+# crypto-wasm
+
+WASM bindings covering the `crypto-wasm` crate. On-chain crypto primitives are still in the `wasm` crate.
